[kv] Avoid duplicate bucket key encoding for Lookup and PrefixLookup (#317)
This also makes Lookuper and PrefixLookuper to initialize and validate for themselves. This makes logic much clean, and avoid mis-reuse for some variables (e.g., lakeTableBucketAssigner).

This also introduced a UnifiedLookuper in Flink connector to unify the different API of Lookuper and PrefixLookup. This makes the logic of AsyncLookupFunction and LookupFunction simpler.